Move mkdir_p function to utils module
